move standalone struct def to standalone.go
